epic: rename from epic-eic
